pruebas de integracio ok

+++ src/main/java/system/pack/converter/TeacherConverter.java
@@ -28,8 +28,8 @@ public class TeacherConverter {
 		teacherBean.setDoctorDegree(teacherEntity.getDoctorDegree());
 		teacherBean.setInstitutionalMail(teacherEntity.getInstitutionalMail());
 		teacherBean.setPersonalMail(teacherEntity.getPersonalMail());
-		teacherBean.setCellNumber(Integer.toString(teacherEntity.getCellNumber()));
-		teacherBean.setHomeNumber(Integer.toString(teacherEntity.getHomeNumber()));
+		teacherBean.setCellNumber(Long.toString(teacherEntity.getCellNumber()));
+		teacherBean.setHomeNumber(Long.toString(teacherEntity.getHomeNumber()));
 		teacherBean.setExperience(teacherEntity.getExperience());
 
 		
@@ -50,8 +50,8 @@ public class TeacherConverter {
 		teacherEntity.setDoctorDegree(teacherBean.getDoctorDegree());
 		teacherEntity.setInstitutionalMail(teacherBean.getInstitutionalMail());
 		teacherEntity.setPersonalMail(teacherBean.getPersonalMail());
-		teacherEntity.setCellNumber(teacherBean.getCellNumber().equals("") ?  0 : Integer.parseInt(teacherBean.getCellNumber()));
-		teacherEntity.setHomeNumber(teacherBean.getCellNumber().equals("") ?  0 : Integer.parseInt(teacherBean.getHomeNumber()));
+		teacherEntity.setCellNumber(teacherBean.getCellNumber().equals("") ?  0 : Long.parseLong(teacherBean.getCellNumber()));
+		teacherEntity.setHomeNumber(teacherBean.getCellNumber().equals("") ?  0 : Long.parseLong(teacherBean.getHomeNumber()));
 		teacherEntity.setExperience(teacherBean.getExperience());
 		
 		return teacherEntity;
